// src/app/demo/models/modelsgeneral/paisviewmodel.ts
export class Pais{
    pais_Id?: number;
    pais_Nombre?: string;
    pais_Codigo?: string;
    pais_Prefijo?: string;
    usua_Creacion?: number;
    usua_Modificacion?: number;
    usuarioCreacion?:string;
    usuarioModificacion?:string;
    pais_FechaCreacion?: string;
    pais_FechaModificacion?: string;

    codigo?: string;
}